create Hooks.ts file to store custom hooks
- Add Hooks.ts
- Move useCustomersContext hook to Hooks file
- Export CustomerContext to allow using it in Hooks file
