+++ src/store/CustomerStore.tsx
@@ -1,4 +1,4 @@
-import { FC, createContext, useCallback, useContext, useState } from "react";
+import { FC, createContext, useCallback, useState } from "react";
 
 import {
   CustomerInterface,
@@ -7,9 +7,9 @@ import {
 } from "../utils/Interface";
 
 // Create Customer Context
-const CustomersContext = createContext<CustomersContextInterface | undefined>(
-  undefined
-);
+export const CustomersContext = createContext<
+  CustomersContextInterface | undefined
+>(undefined);
 
 export const CustomersProvider: FC<ChildrenInterface> = ({ children }) => {
   // store customers data in customersList
@@ -25,15 +25,4 @@ export const CustomersProvider: FC<ChildrenInterface> = ({ children }) => {
       {children}
     </CustomersContext.Provider>
   );
-};
-
-// create custom hook to facilitate usage of the context and handle error of using context values
-// (properties and methods ) without provider
-export const useCustomersContext = () => {
-  const customersContext = useContext(CustomersContext);
-
-  if (!customersContext)
-    throw new Error("You need to use this context inside CustomerProvider");
-
-  return customersContext;
 };
